requirements.txt
Flask==2.2.5
gunicorn==21.2.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
PyJWT==2.8.0
APScheduler==3.10.4
Jinja2==3.1.2